requirements-test.txt
# Testing dependencies for SupplyGraph
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
websockets>=12.0
psycopg2-binary>=2.9.0
redis>=5.0.0
prisma>=0.15.0
python-dotenv>=1.0.0
pytest-mock>=3.12.0
pytest-cov>=4.1.0
pytest-xdist>=3.3.1
asyncio>=3.4.3
coverage>=7.3.0